feat: Add Sabha component
Add a new component called Sabha to the app/(tabs) directory. This component renders a View with a Text component displaying the text "sabha". This component will be used in the app's tab layout.

Refactor Header component

In the Header component, update the className of a specific element to "pr-2 flex flex-row gap-1 items-center" instead of "pr-2 flex flex-col gap-1 items-center". This change modifies the layout of the element.

Update icons in constants/indxe.ts

In the constants/indxe.ts file, add new icons for azkar, kaaba, salat, and sabha. These icons are imported from the respective image files and added to the icons object.

Refactor [category] component

In the [category] component of the azkar directory, remove the import statement for SafeAreaView from react-native-safe-area-context and add it from react-native. This change ensures that the correct SafeAreaView component is used.

Update __routes in .expo/types/router.d.ts

In the .expo/types/router.d.ts file, update the __routes interface to include a new route for "/(tabs)/sabha". This change reflects the addition of the Sabha component to the app's tab layout.

Update button and input components

In the button component, add a new variant called "white" with the corresponding background and text colors. In the input component, add a conditional className to set the text alignment to right when the i18n language is set to "ar".

Refactor _layout.tsx

In the _layout.tsx file, update the initialRouteName of the Tabs component to "azkar". This change sets the initial active tab to the Azkar tab. Also, update the tabBarStyle to include paddingTop and remove position: "absolute" to improve the layout of the tab bar. Finally, update the TabIcon component to display the title of each tab.

Refactor (tabs)/_layout.tsx

In the (tabs)/_layout.tsx file, add a new tab for the Sabha component. This tab has the title "السبحة" and displays the Sabha icon. The header is hidden for this tab.